# fix: TypeScript performance of `BarPercent.tsx`