--- a/packages/manager/.changeset/pr-11076-tech-stories-1728498604542.md
+++ b/packages/manager/.changeset/pr-11076-tech-stories-1728498604542.md
@@ -1,0 +1,5 @@
+---
+"@linode/manager": Tech Stories
+---
+
+Fix TypeScript Performance of `BarPercent.tsx` ([#11076](https://github.com/linode/manager/pull/11076))